new build changes

+++ dashboard/dep.sh
@@ -4,13 +4,27 @@
 mkdir dashboard
 cp -R WebContent/* dashboard
 tar -zcf dashboard.tgz dashboard
+
+echo "----------------- dashboard"
+echo "Placing tar file in ~/deploy"
+
 mv dashboard.tgz ~/deploy
 
 # copy files to the local server
-rm -rf /Library/WebServer/Documents/dashboard
-cp -R dashboard /Library/WebServer/Documents
 
-rm -rf dashboard
+docdir=$WEBSITE_DOCS/app/dashboard
+
+echo "Website: $WEBSITE_DOCS"
+echo "Deploying to: $docdir"
+
+sudo rm -rf $docdir
+sudo mkdir $docdir
+sudo cp -R dashboard $docdir
+
+if [ -f dep.sh ]
+then
+	rm -rf dashboard
+fi
